Travis CI: use before_script and split commands
The sequence is before_install -> install -> before_script -> script.
`install` is supposed to be the one where dependencies are installed,
but apparently the apt and homebrew plugins do it earlier. I'm not sure
why the docs suggest putting `apt-get install` in `before_install`,
but we no longer even do that. It definitely seems like the wrong
stage for `./autogen.sh` and `./configure`, even if it currently
affects nothing in practice.